+++ models/role.mjs
@@ -199,6 +199,7 @@ export const ROLE_SUPERADMIN = new RoleList(
 
   rights.CAN_LIST_RATING,
   rights.CAN_GET_RATING,
+  rights.CAN_GET_STATS,
 );
 roles.ROLE_SUPERADMIN = ROLE_SUPERADMIN;